FIX:maintain backwards compat semantics with CreateDataset
